+++ NEWS.md
@@ -7,6 +7,8 @@ Red Panda C++ Version 1.0.8
   - enhancement: auto set output suffix to ".s" when compiler commandline argument for "-S" is turned on
   - enhancement: show error message when user set a shortcut that's already being used.
   - enhancement: adjust scheme colors for "dark" and "high contrast" themes
+  - enhancement: can debug files that has non-ascii chars in its path and is compiled by clang
+  - fix: when debugging project, default compiler set is wrongly used
 
 Red Panda C++ Version 1.0.7
   - change: use Shift+Enter to break line